fix: [CDS-34063]: jest
