+++ src/modules/72-templates-library/components/TemplateStudio/StepTemplateCanvas/StepTemplateDiagram/__tests__/StepTemplateDiagram.test.tsx
