+++ src/modules/72-templates-library/components/TemplateStudio/TemplateVariables/__tests__/metadataMap.json
